+++ countlyCommon/TestingRelated/RequestTestCases.cs
@@ -1,13 +1,15 @@
-﻿using CountlySDK.CountlyCommon.Entities;
-using CountlySDK.CountlyCommon.Helpers;
-using CountlySDK.Entities.EntityBase;
-using CountlySDK.Helpers;
-using System;
+﻿using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
+using System.Collections.Specialized;
 using System.Threading.Tasks;
+using System.Web;
+using CountlySDK;
+using CountlySDK.CountlyCommon.Entities;
+using CountlySDK.CountlyCommon.Helpers;
+using CountlySDK.Entities;
+using CountlySDK.Entities.EntityBase;
 using Xunit;
+using static CountlySDK.CountlyCommon.Helpers.RequestHelper;
 using static CountlySDK.Helpers.TimeHelper;
 
 namespace TestProject_common
@@ -20,6 +22,9 @@ namespace TestProject_common
         public RequestTestCases()
         {
             TestHelper.CleanDataFiles();
+            Countly.Halt();
+            TestHelper.CleanDataFiles();
+            Countly.Instance.deferUpload = true;
         }
 
         /// <summary>
@@ -30,50 +35,104 @@ namespace TestProject_common
         }
 
         [Fact]
-        public void BaseRequestBasic()
+        /// <summary>
+        /// It validates base request parameters.
+        /// </summary>
+        public async void ValidateBaseRequestParams()
         {
-            TimeHelper timeHelper = new TimeHelper();
-            TimeInstant instant = timeHelper.GetUniqueInstant();
-            DeviceId dId = new DeviceId("b", DeviceBase.DeviceIdMethodInternal.developerSupplied);
-            string req = RequestHelper.CreateBaseRequest("a", dId, "c", "d", instant);
-            string expected = string.Format("/i?app_key={0}&device_id={1}&timestamp={2}&sdk_version={3}&sdk_name={4}&hour={5}&dow={6}&tz={7}&t={8}", "a", "b", instant.Timestamp, "c", "d", instant.Hour, instant.Dow, instant.Timezone, dId.Type());
 
-            Assert.Contains(expected, req);
+            CountlyConfig cc = new CountlyConfig() {
+                serverUrl = "https://try.count.ly/",
+                appKey = "YOUR_APP_KEY",
+                developerProvidedDeviceId = "test device id",
+                deviceIdMethod = Countly.DeviceIdMethod.developerSupplied
+            };
+
+            await Countly.Instance.Init(cc);
+            Dictionary<string, object> baseParams = await Countly.Instance.requestHelper.GetBaseParams();
+
+            Assert.Equal(9, baseParams.Count);
+
+            Assert.Equal("YOUR_APP_KEY", baseParams["app_key"]);
+            Assert.Equal("test device id", baseParams["device_id"]);
+            Assert.Equal("22.02.1", baseParams["sdk_version"]);
+            Assert.Equal(0, baseParams["t"]);
+
+            Assert.True(baseParams.ContainsKey("sdk_name"));
+            Assert.True(baseParams.ContainsKey("timestamp"));
+            Assert.True(baseParams.ContainsKey("dow"));
+            Assert.True(baseParams.ContainsKey("hour"));
+            Assert.True(baseParams.ContainsKey("tz"));
         }
 
         [Fact]
-        public void LocationRequestBasic()
+        /// <summary>
+        /// It validates request builder.
+        /// </summary>
+        public async void ValidateRequestBuilder()
         {
-            string res2 = RequestHelper.CreateLocationRequest("asd");
-            Assert.Null(res2);
+            Dictionary<string, object> param = new Dictionary<string, object>
+             {
+                {"a", "A"},
+                {"b", "B"},
+                {"1", 1},
+                {"2", true},
+            };
 
-            string res3 = RequestHelper.CreateLocationRequest("asd", null, null, null, null);
-            Assert.Null(res3);
+
+            RequestHelper requestHelper = new RequestHelper(new IRequesttHelperImpl());
+
+            string request = await requestHelper.BuildRequest(param);
+
+            NameValueCollection collection = HttpUtility.ParseQueryString(request.Substring(2));
+
+            Assert.Equal("A", collection.Get("a"));
+            Assert.Equal("B", collection.Get("b"));
+            Assert.Equal("1", collection.Get("1"));
+            Assert.Equal("True", collection.Get("2"));
+
+            Assert.Equal("app-key", collection.Get("app_key"));
+            Assert.Equal("0", collection.Get("t"));
+            Assert.Equal("sdk-name", collection.Get("sdk_name"));
+            Assert.Equal("sdk-version", collection.Get("sdk_version"));
+            Assert.Equal("device-id", collection.Get("device_id"));
+
+            Assert.Equal("300", collection.Get("tz"));
+            Assert.Equal("2", collection.Get("dow"));
+            Assert.Equal("7", collection.Get("hour"));
+            Assert.Equal("1666683640551", collection.Get("timestamp"));
+
         }
 
-        [Fact]
-        public void LocationRequestSimple()
+        private class IRequesttHelperImpl : IRequestHelper
         {
-            string br = "asd";
-            string res;
-            res = RequestHelper.CreateLocationRequest(br, null, null, null, null);
-            Assert.Null(res);
+            public string GetAppKey()
+            {
+                return "app-key";
+            }
 
-            res = RequestHelper.CreateLocationRequest(br, "", "", "", "");
-            Assert.Equal("asd&location=&ip=&country_code=&city=", res);
+            public async Task<DeviceId> GetDeviceId()
+            {
+                DeviceId deviceId = new DeviceId("device-id", DeviceBase.DeviceIdMethodInternal.developerSupplied);
+                return deviceId;
+            }
 
-            res = RequestHelper.CreateLocationRequest(br, null, "a", "b", "c");
-            Assert.Equal("asd&ip=a&country_code=b&city=c", res);
+            public string GetSDKName()
+            {
+                return "sdk-name";
+            }
 
-            res = RequestHelper.CreateLocationRequest(br, "a", null, "b", "c");
-            Assert.Equal("asd&location=a&country_code=b&city=c", res);
+            public string GetSDKVersion()
+            {
+                return "sdk-version";
+            }
 
-            res = RequestHelper.CreateLocationRequest(br, "a", "b", null, "c");
-            Assert.Equal("asd&location=a&ip=b&city=c", res);
-
-            res = RequestHelper.CreateLocationRequest(br, "a", "b", "c", null);
-            Assert.Equal("asd&location=a&ip=b&country_code=c", res);
+            public TimeInstant GetTimeInstant()
+            {
+                long timestamp = 1666683640551;
+                TimeInstant timeInstant = TimeInstant.Get(timestamp);
+                return timeInstant;
+            }
         }
-
     }
 }